Desktop/FakeNews/ml_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
scikit-learn==1.4.2
joblib==1.4.2
numpy==1.26.4
httpx==0.27.0
python-dotenv==1.0.1
sentence-transformers==2.7.0
faiss-cpu==1.8.0
openai==1.30.1
groq==0.4.1
